Use Shared Memory in Gaussian_Blur
Got shared image tile and filter to work - Did not work if the block size was too small (16*8) which was fastest without shared memory usage.  Works fine with 16*16 block size and 20*20, but would not work when block got larger e.g. 32*32 (which also worked fine when not using shared memory).

Could not get it to work when applying halo piece by piece on 4 edges and corners individually (looks like something is not correct on an edge case at a corner).  Worked fine when using 4 stamps of entire block which are offset to reach each corner.  

Set up defines to show different cases.